+++ manual/android-client/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Fixed
 
+* Show bottom toolbar when viewing specific history lists.
+  ([#75](https://j.njsm.de/git/veenj/stocks/issues/75)).
+
 * Locations are now ordered alphabetically when adding food items.
   ([#75](https://j.njsm.de/git/veenj/stocks/issues/75)).